Add minor fixes
Includes some reorganization of get_ci_seq to avoid circular
dependencies, explicitly defining some predmix *_ci_seq functions, and
other bug fixes.
